refactor js
