+++ srcs/frontend/index.html
@@ -4,8 +4,9 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>BeePong</title>
-    <script src="/js/form.js"></script>
-    <script src="/js/route.js"></script>
+    <script type="module" src="/js/main.js"></script>
+    <script type="module" src="/js/form.js"></script>
+    <script type="module" src="/js/route.js"></script>
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Krona+One&display=swap" rel="stylesheet">
@@ -24,7 +25,7 @@
 <body>
     <nav>
         <div>
-            <a href="/" onclick="navigate(event)">
+            <a href="/" >
                 <img src="/img/beePong.svg" alt="beePong logo"/>
             </a>
         </div>
