+++ srcs/frontend/js/form.js
@@ -1,32 +1,14 @@
-//handle form submit for register, login and logout
+import { loadNavBar } from "./navbar.js";
+
+// Handle form submission
 async function handleFormSubmit(event) {
 	event.preventDefault();
+
 	const form = event.target;
-	// Create a FormData object from the form element, which captures all form input values
-	const formData = new FormData(form);
-
-	// Initialize a URLSearchParams object to handle URL-encoded data
-	const data = new URLSearchParams();
-	// Iterate over each key-value pair in the FormData object
-	// Append each key-value pair to the URLSearchParams object
-	formData.forEach((value, key) => {
-		data.append(key, value);
-	});
-
-	// Create a new URL object from the form's action attribute, which provides the URL for form submission
 	const url = new URL(form.action);
-	// Combine the URL's pathname (the path part of the URL) with its search parameters (query string)
-	// This results in the relative path of the form's action URL, including any query parameters
-	const relativePath = url.pathname + url.search;
 
 	try {
-		const response = await fetch(`/form${relativePath}`, {
-				method: 'POST',
-				headers: {
-					'Content-Type': 'application/x-www-form-urlencoded'
-				},
-				body: data.toString()
-		});
+		const response = await submitForm(form, url);
 
 		if (!response.ok) {
 			if (response.status !== 400)
@@ -35,37 +17,61 @@ async function handleFormSubmit(event) {
 
 		const result = await response.json();
 
-		if (result.success) {
-				// Handle redirection
-				if (relativePath === '/accounts/logout/')
-					navigate('/');
-				else
-					navigate(result.redirect);
+		if (result.success) { // Handle redirection
+			navigate(result.redirect);
+			if (url.pathname === '/accounts/login/' || url.pathname === '/accounts/logout/')
 				loadNavBar();
 		}
-		else {
-				const errorContainer = document.querySelector('.form-error-message');
-				// Clear any previous errors
-				errorContainer.innerHTML = '';
-
-				// Display the error message(s)
-				if (result.errors) {
-						// Show the error container
-						errorContainer.style.display = 'block';
-
-						// Iterate through the errors object to display each error message
-						for (const key in result.errors) {
-								if (result.errors.hasOwnProperty(key)) {
-										result.errors[key].forEach(error => {
-												const errorElement = document.createElement('p');
-												errorElement.textContent = error;
-												errorContainer.appendChild(errorElement);
-										});
-								}
-						}
-				}
+		else { // Display error messages
+			displayFormErrors(result.errors);
 		}
+
 	} catch (error) {
 			console.error('There was a problem with the fetch operation:', error);
 	}
 }
+
+// Submit the form and return the response
+async function submitForm(form, url) {
+	const urlEncodedData = getFormDataAsUrlEncoded(form);
+	return await fetch(`/form${url.pathname}`, {
+		method: 'POST',
+		headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
+		body: urlEncodedData.toString(),
+	});
+}
+
+// Extract data from a form and convert it to URL-encoded format
+function getFormDataAsUrlEncoded(form) {
+	const formData = new FormData(form);
+	const urlEncodedData = new URLSearchParams();
+
+	formData.forEach((value, key) => { // Append each key-value pair in the FormData object to the URLSearchParams object
+		urlEncodedData.append(key, value);
+	});
+
+	return urlEncodedData;
+}
+
+// Display error messages in the form
+function displayFormErrors(errors) {
+	const errorContainer = document.querySelector('.form-error-message');
+	errorContainer.innerHTML = ''; // Clear previous errors
+
+	if (errors) {
+		errorContainer.style.display = 'block'; // Show the error container
+		
+		for (const key in errors) {
+			if (errors.hasOwnProperty(key)) {
+				errors[key].forEach(error => {
+					const errorElement = document.createElement('p');
+					errorElement.textContent = error;
+					errorContainer.appendChild(errorElement);
+				});
+			}
+		}
+	}
+}
+
+// Attach handleFormSubmit to the global window object for use in inline event handlers
+window.handleFormSubmit = handleFormSubmit;
